+++ R/print.summary.meta.R
@@ -12,8 +12,6 @@
 #'   meta-analysis should be conducted.
 #' @param random A logical indicating whether a random effects
 #'   meta-analysis should be conducted.
-#' @param prediction A logical indicating whether a prediction
-#'   interval should be printed.
 #' @param details A logical indicating whether further details of
 #'   individual studies should be printed.
 #' @param ma A logical indicating whether the summary results of the
@@ -21,11 +19,6 @@
 #' @param overall A logical indicating whether overall summaries
 #'   should be reported. This argument is useful in a meta-analysis
 #'   with subgroups if overall results should not be reported.
-#' @param overall.hetstat A logical value indicating whether to print
-#'   heterogeneity measures for overall treatment comparisons. This
-#'   argument is useful in a meta-analysis with subgroups if
-#'   heterogeneity statistics should only be printed on subgroup
-#'   level.
 #' @param backtransf A logical indicating whether printed results
 #'   should be back transformed. If \code{backtransf = TRUE}, results
 #'   for \code{sm = "OR"} are printed as odds ratios rather than log
@@ -147,10 +140,8 @@ print.summary.meta <- function(x,
                                sortvar,
                                fixed = x$x$fixed,
                                random = x$x$random,
-                               prediction = x$prediction,
                                details = FALSE, ma = TRUE,
                                overall = x$overall,
-                               overall.hetstat = x$overall.hetstat,
                                ##
                                backtransf = x$backtransf,
                                pscale = x$pscale,
@@ -222,13 +213,10 @@ print.summary.meta <- function(x,
   if (!sort)
     sortvar <- 1:k.all
   ##
-  chklogical(prediction)
   chklogical(details)
   chklogical(ma)
   overall <- replaceNULL(overall, TRUE)
   chklogical(overall)
-  overall.hetstat <- replaceNULL(overall.hetstat, TRUE)
-  chklogical(overall.hetstat)
   ##
   if (is.untransformed(x$sm))
     backtransf <- TRUE
@@ -275,9 +263,6 @@ print.summary.meta <- function(x,
   chkchar(text.tau2, length = 1)
   chkchar(text.tau, length = 1)
   chkchar(text.I2, length = 1)
-  tt2 <- text.tau2
-  tt <- text.tau
-  ti <- text.I2
   ##
   ## Catch 'truncate' from meta-analysis object:
   ##
@@ -348,10 +333,6 @@ print.summary.meta <- function(x,
   ##
   metainf.metacum <- inherits(x, "metainf") | inherits(x, "metacum")
   mb.glmm <- inherits(x, "metabind") | x$method == "GLMM"
-  ##
-  prediction <- prediction & x$k >= 3
-  if (is.na(prediction))
-    prediction <- FALSE
   ##
   ci.lab <- paste0(round(100 * level, 1), "%-CI")
   ##
@@ -876,8 +857,7 @@ print.summary.meta <- function(x,
                  header = FALSE,
                  digits = digits,
                  fixed = fixed, random = random,
-                 prediction = prediction,
-                 overall = overall, overall.hetstat = overall.hetstat,
+                 overall = overall,
                  backtransf = backtransf, pscale = pscale,
                  irscale = irscale, irunit = irunit,
                  digits.tau2 = digits.tau2, digits.tau = digits.tau,